validate ci.yml file
Add some basic test checks for the new ci.yml to ensure all job
dependencies are set up right.

We can expand this with more checks later.

(cherry picked from commit 66e7c0607353c0ddc0ced3b57d7c030417e9239f)
